División en Editor Para simplificar el archivo "editor" se ha decidido modularizarlo en otros más pequeños.
