+++ func/util/elementos.js
@@ -5,3 +5,12 @@ export function crear_botón(id, contenido) {
 	botón.id = id;
 	return botón;
 }
+
+export function en_años(fecha) {
+	let tiempo = fecha.año;
+	if (fecha.mes)
+		tiempo += Number(fecha.mes / 12);
+	if (fecha.día)
+		tiempo += Number(fecha.día / 365);
+	return tiempo;
+}
